Update secure_headers from 3.7.3 to 6.0.0
Previous versions of `secure_headers` obscured a bug that should have
caused the spec on line 79 of `spec/features/users/sign_up_spec.rb` to
fail. Before version 6.0.0, the gem kept track of which CSP directives
weren't supported in which browsers, and parsed the user agent and
overrode the app's CSP config to remove any directives that it thought
were not supported.

With the user agent parsing code, the `HeadlessChrome` User Agent was
being interpreted by the `secure_headers` gem as a browser that does not
support the `form-action` directive, which allowed the aforementioned
spec to pass (because the `form-action` directive was removed from the
CSP headers). I verified this using the middleware linked to in this
[blog post](https://about.gitlab.com/2017/12/19/moving-to-headless-chrome/),
which allows us to inspect the response headers (Selenium doesn't have
support for `page.response_headers`).

In the latest version of `secure_headers`, the `form-action` directive
is not removed (which is a good thing), and the reason the test fails is
because this is a JS test, and we have configured all JS tests to use
`127.0.0.1` as the domain name, but in the last part of the test, when
the user clicks on the button in the modal to cancel account creation
and delete their account, it tries to go to
`http://www.example.com/sign_up/start`.

This is because the `UsersController` determines where to redirect based
on the `cancel_link_url` of the `decorated_session`, which in this case
is the `ServiceProviderDecoratedSession`, whose `cancel_link_url` was
constructed using `Rails.application.routes.url_helpers`, which, for
some reason, ignores the `default_url_options` defined in
`ApplicationController` and defaults the host to `www.example.com`, and
since the `form-action` directive doesn't allow `example.com`, the test
fails to redirect. Note that the defaulting to `www.example.com` only
happens in the test environment. I verified this works fine in
production.

I verified this by replacing `redirect_to url_after_cancellation` with
`redirect_to sign_up_start_url` in `UsersController`, which allowed the
test to pass. So then I tried passing in the `host` parameter to
`sign_up_start_url` in the decorator, but that didn't work because
Capybara uses a port with `127.0.0.1`, but the `sign_up_start_url` in
the decorator doesn't include the port, and so the `form-action`
directive will not allow this redirect. The port is random each time the
test is run, so there's no obvious way to set the port, and
`Capybara.always_include_port = false` didn't work.

So then I looked at the decorator class some more, and noticed the
`view_context` that gets passed to it and it struck me: we should be
using `view_context.sign_up_start_url` instead of including
`Rails.application.routes.url_helpers`, and 🎉!