[LG 240] piv/cac is only for certain agencies
**Why**:
Piv/cac requires that we have a set of signing
certificates to validate presented certs. This
requires research. We don't have them for all
of the federal government, so we want to keep
from confusing almost everyone.

**How**:
Only show the piv/cac option if the user has
an identity tied to the right preconfigured
agencies.