[LG-272] Support randomizable Piv/Cac service URLs
**Why**:
Browsers cache information about which certificate was
used with which hostname. If a bad certificate is selected,
the only recourse is to shut down the browser and restart it.
This leads to a bad UX.

**How**:
We have wildcard hostnames for the piv/cac service, so we
generate a random hostname each time we redirect the browser.
We try to balance between randomness and DNS caching.