[LG-267] Auth requests to decode pivcac tokens
**Why**:
We don't want everyone being able to take a token and
inspect it. We only want the IdP able to do this.

**How**:
We have several options, including client/server TLS
certificates. But adding an HMAC with a shared secret
is the lightest weight way to lock down the decoding
endpoint.