Don't set up 2fa without authenticating first
**Why**:
If someone creates an account and adds their piv/cac
and then logs out, on subsequent login, they had the
option of adding a phone number when prompted for
their piv/cac. This allowed someone with the username
and password to bypass the piv/cac second factor and
set the phone, which would then be used as the
second factor.

**How**:
Only allow phone setup during account creation or after
authenticating with a second factor.