[LG-256] Add spinner when presenting piv/cac cert
**Why**:
The browser UX might take a while, and we don't want the
user to feel that everything is frozen up.

**How**:
When the user clicks on the button to present their piv/cac
cert, we intercept the event and change the content to be a
GIF spinner. We let the event continue so the browser follows
through with the request.