LG-319 Record password strength stats
**Why**: So that we can measure how our password policies are affecting
the stength of passwords people are creating

**How**: We do not want the password metrics to be associated with the
user records, so we create a new record, named `PasswordMetric` which
captures the name of the metric we are calculating, value (e.g. the
length of the password), and count (e.g. how many users have a password
of a given length). The code then does an atomic insert any time a user
sets, resets, or changes their password.