LG-459 Clean up localizations
**Why**:
- We like to keep our keys sorted alphabetically, but many files were
not sorted
- The configuration for the `i18n-tasks` gem was ignoring all strings
in the `config/locales/devise` folder, and so we ended up with lots of
unused strings.

**How**:
- Run `i18n-tasks normalize` to sort the keys
- Add `i18n-tasks normalize` to the `make normalize_yaml` script so
that keys get sorted as part of our normalization
- Remove the `line-width` setting from the `i18n-tasks` config because
when you run `i18n-tasks normalize`, it also changes the formatting to
not wrap at 80 characters, but we expect our files to be wrapped at
80 characters (which is the default behavior of Psych, the Ruby YAML
parser).
- Move the non-Devise strings to their own section and ignore them
when `i18n-tasks` thinks they are unused (usually due to interpolation)
- Remove any remaining unused files using `i18n-tasks remove-unused`