Prioritize piv/cac during signin if configured
**Why**:
If a user has configured their account with piv/cac, then
we want to show that rather than sending them a text. They
can still opt to receive a text if they aren't able to use
their piv/cac.

**How**:
Swap the ordering so piv/cac is ahead of sms/voice.