Revert VoiceOtpSenderJob#send_otp
**Why**:
A bug was found in RC 61, but shortly after the release branch was
created, #2276 was merged into master with changes incompatible with
RC 61. When I fixed the bug in RC 61, I didn't notice VoiceOtpSenderJob
had changed. This PR uses the old code for #send_otp to allow tests in
RC 61 to pass.