LG-268 Serve assets from Cloudfront CDN
**Why**: Moving assets to be served from Cloudfront as opposed to nginx
will reduce the load on our servers, and will make the site faster for
users across the globe.

On the Rails app side, this is a trivial change. The tricky part is
setting up the Cloudfront distribution and making sure to select
`Whitelist` from the `Cache Based on Selected Request Headers` dropdown,
and add the `Origin` header to the `Whitelist Headers`. Then, add the
following headers to the nginx config for fonts:
`Access-Control-Allow-Origin` set to the current server domain name,
including the protocol, such as `https://secure.login.gov`, and
`Access-Control-Allow-Methods` set to `GET`.