LG-439 Don't raise error for invalid user params
**Why**: There is a bug in Devise where if you visit the sign in page
with a `user` param set to a String, it will raise an exception. The
bug has been fixed and merged into master, but hasn't been released
to Rubygems yet. We could either point our gem to Devise's master
branch, or temporarily add some code to prevent the bug until the fix
is released. I opted for the latter because I don't know how stable the
master branch is and what kind of changes are in it.