Stop reading personal key columns (#2313)
**Why**: We're phasing out the old personal key columns in favor of
using `encrypted_recovery_code_digest`.

The old code was actually using the `#personal_key` method in a lot of
places to determine if a user had a personal key, so this code switches
that up. It also refactors the personal key to behave a bit more like a
password. That could lay the groundwork for some more refactoring, but I
didn't want to pursue it too far until @jgsmith-usds gets the 2FA
refactor done.