LG-433 Add script to email compromised users
**Why**: If we get a report that a user's password might have been
compromised, we need a way to quickly reset the user's password and send
them an email prompting them to reset their password.

**How**:
- Create an email template with localizable content
- Create a rake task that accepts a comma-separated list of user emails,
and iterates through each email, and if it finds a corresponding user,
will reset the user's password to a random value, then will email the
user to let them know and that they need to reset their password to
be able to sign back in.