LG-490 Read password hashes without digested salts
**Why**: Previously, we generated 20 char random strings for our
password salts. To get those string to an appropriate length for an
SCrypt salt, we took a SHA256 digest of the string.

We plan to start using 32 random bytes for the salt so we can use them
directly instead of hashing them. This commit adds the code to read the
new digests and tests to verify the old digests can still be verified
after the change to 32 random bytes is implemented.

This commit will not have any impact when deployed, but allows instances
generating 20 byte and 32 byte digests to serve traffic together so that
we can make the change to 32 byte digests.