LG-491 Don't truncate scrypt digest in pii encryption
**Why**: The original 2L-KMS implementation truncated the output from
the scrypt digest. We wanted to configure SCrypt to output 32 bytes, but
after checking, it was already configured to do that. The reason we had
to truncate, is because the output was hex encoded making it 64
characters long.

This commit hex decodes the digest and uses the 32 byte result to
encrypt the PII.