[LG-498] Add phone configurations table
**Why**:
We want to support multiple phone configurations for a user.

**How**:
Move the phone configuration from the users table to a new table.
This PR is step 1 of several: create the table and write to it.

The critical work here is to make sure that every time we update
the user's 2fa phone, we also update or write to the new table.