Fix error in IdV flow when USPS is disabled
**Why**: The OTP confirmation page had a call to `idv_usps_path` which
was not behind a conditional for whether USPS verification was disabled
for the app. This meant that a path helper for a route that does not
exist was called. This raised and caused a 500.

This commit puts the path method behind a conditional and adds a feature
spec that goes through IdV with USPS disabled to make sure this doesn't
happen in the future.