[LG-509] Only allow expected hosts in requests
**Why**:
It is possible to modify the Host: header in a request and
have that header be reflected in the link that devise uses
to direct people to change a forgotten password. We don't
want that.

**How**:
Check that the request is being made to an expected host.
If not, redirect to the canonical host.