LG-544 Prevent calling unsupported countries
**Why**: Due to a bug that has since been fixed, there still exist
users in the database whose `otp_delivery_preference` is set to `voice`
even though our Twilio account does not allow us to call their phone
number. This means that every time they sign in, we try to make a call,
which results in a Twilio error, then the user has to choose a different
2FA option.

I had noticed this scenario a while back and fixed it, when the country
in question was India, but this time, people in Bermuda are experiencing
this. The reason why it was failing for Bermuda is because it uses the
same `+1` dialing code as the US, and we were treating all countries
that use `+1` as having "area codes", and we were trying to match the
area code by calling the `area_code` method on the parsed phone. It's
possible that this was working with `phony_rails`, but it doesn't with
`Phonelib` (which is accurate).

**How**: Given that we have a list of all countries in our YAML file
and in the UI dropdown, and that Phonelib knows which country the number
comes from, we don't need to treat countries with `+1` differently. This
allows us to remove all code related to unsupported area codes.

Instead of looking for an area code, we look up the 2-letter code from
the parsed phone, and find a match in our YAML file.