Track additional IdV analytics
**Why**: To help us better understand how people use the product and
where they are failing.

The main change here is to capture the controller where the user decided
to cancel verification. This will allow us to see if there's a
particular blocker in the flow. This information is used by other
controllers, so I refactored a bit to reduce duplication.

I also added a visit event for the come back later page, and made sure
that the call to track the visit event in the IdV Sessions Controller
was the first thing in the method. Otherwise, if an error happens before
the analytics call, it wouldn't be easy to tell whether or not the user
visited (or was redirected to) the sessions controller.

For troubleshooting, it's very helpful to have events for all page
visits.