[LG-487] piv/cac available based on email domain
**Why**:
We want to roll out piv/cac across the government, but in a
measured way since we don't want to advertise it to folk we
know that we can't support their piv/cac.

**How**:
Allow a configurable list of domains that we know we should
be able to support. Allow suffix matching if starting with a
`.` so we can support an agency that uses multiple subdomains,
but also allow full matching for those agencies for which we
know we can only support a subset. We assume that a subset of
an agency for our purposes here will share an email domain.

We don't require a user to have a proper email *and* a
relationship with a particular SP. We require at least one
of a relationship with a particular SP and a supported
email domain.