LG-559 Allow sign in via remember me after idling
**Why**: I found a bug that I can consistently reproduce in Chrome (but
not in other browsers for some reason):
If you already have an identity associated with an SP, and are using the
"remember me" feature, you won't be able to sign in during this
scenario:
1. Visit the SP
2. Click through to the IdP
3. Click "Sign in" but don't sign in yet
4. Wait 15 minutes, or until the page refreshes and says "For your
security, we clear what you entered..."
5. Enter your email and password and submit the form

Result: Chrome will prevent the redirect back to the SP because the
redirect URI is not in the CSP headers.

**How**: Store the SP info in the session when the sign in page is
visited because the CSP headers only get updated if the SP info is in
the session, but if you stay idle for 15 minutes, the session gets
deleted from Redis.