Fix a flickering spec in the OTP rate limiter (#2448)
**Why**: The OTP rate limiter was doing an exact match on
`second_factor_locked_out` after rate limitting a user. If the system
clock advances to the next second between the user being locked out and
the check against the current time, this test fails. This happens
frequently on CI. This commit uses rspec's `be_within.of` matcher
instead of using an exact match.