Copy phone info to new table sans crypto
**Why**:
Each time we decrypt/encrypt in AWS, we call KMS.
Since we don't really need to know the content of
the phone number to do the copy, we don't need to
risk running into KMS limits during the copy.

**How**:
We copy the encrypted data rather than going through
the crypto round-trip.