Remove proofer initializer (#2447)
**Why**: Autoloader was reloading the `Idv::Proofer` class when there
were changes to the app. As a result, the configurations set in the
initializer were being wiped out since the initializer was not re-run
when the class was autoloaded.

Since the configurations in the proofer class that were being set are
all around fetching vendors, they could be replaced with code that
lazily instantiates the list of vendors. That is what this PR does.

This PR also has some code to do some refactoring and simplify some of
the logic.