[LG-500] Read phone configuration from table (#2441)
* [LG-500] Read phone configuration from table

**Why**:
We want to transition to the phone_configurations table so
that we can support multiple phones per user. For this PR,
we read phone information from the new table without changing
how we write data.

**How**:
We change everywhere that we read phone info from the user model
to read it from the table.