LG-644 Ensure rack-timeout is properly configured
**Why**: When we last upgraded the rack-timeout gem, we didn't follow
the upgrade instructions carefully enough (my bad). The two options are:

- Let the gem define the middleware, and the only thing needed on our
end is to add the `RACK_TIMEOUT_SERVICE_TIMEOUT` env var if we want to
override the default setting of 15 seconds.

- Define the middleware manually, which requires making sure to insert
it at the proper point.

We had defined it manually, but placed it after all our other
middleware in `config/application.rb`, which meant that it would never
stop long-running requests. The consequence is that we ended up having
some Twilio requests take longer than 15 minutes! The fact that we are
making Twilio requests without defining a timeout is another bug for
which there is a separate JIRA issue.

**How**: Implement the first solution, which is the simplest.