LG-627 Don't automatically select SMS during IdV
**Why**: When a user uses a different number for address verification
than the one they used for 2FA they are asked to choose if they'd like
to get the code sent to them via text message or voice message. Text
message is auto-selected but it should default to no selection so users
don't think text message is the only option or that it's automatically
compatible with their phone.