LG-558 Make jurisdiction step non-re-entrant
**Why**: The jurisdiction step is at the beginning flow to prevent users
who are in an unsupported jurisdiction from having to enter all of their
info. It was re-entrant which meant that anytime you returned to the
beginning of the flow, you had to complete it, even if you had already
verified your info. This doesn't make sense so this commit configures
the step such that you cannot return to the jurisdiction step after
completing it.