LG-651 Allow newrelic_rpm to be updated
**Why**: The `newrelic_rpm` gem added a call to `SecureRandom.hex(8)`
on 6/13/18:
https://github.com/newrelic/rpm/commit/60ef61c512a29107de5015b1d574efc4566d08f6

Some of our tests stub calls to `SecureRandom`, and because we have
New Relic tracers in an initializer, the newly-added call to
`SecureRandom` in the gem caused the tests to fail because they are
not expecting that additional call.

RSpec tells us to stub the call with a default value first in this case.