LG-671 Don't raise if TOTP secret not in session
**Why**: We see half a dozen or so errors in production every month due
to folks submitting the authentication app setup form when the TOTP
secret is no longer in the session. The `two_factor_authentication` gem
raises an error if the secret is nil, but we don't want to raise an
error. We want to display a helpful message to the user so they can try
again.

**How**: Don't call the gem's `confirm_totp_secret` method if the
secret is nil. Instead, treat it as an invalid input.