LG-670 Handle SES timeouts gracefully
**Why**: In a few rare cases, SES does not respond in time when the app
tries to send an email. In the reset password scenario, the controller
was attempting to send an email after it redirected the user, but if
the call to AWS SES takes longer than 15 seconds, RackTimeout kicks in,
and `ApplicationController` calls `render_timeout`, which then causes
the following exception:

```
An AbstractController::DoubleRenderError occurred in reset_passwords#update:

Render and/or redirect were called multiple times in this action.
Please note that you may only call render OR redirect, and at most once
per action. Also note that neither redirect nor render terminate
execution of the action, so if you want to exit an action after
redirecting, you need to do something like "redirect_to(...) and return".
```

The quick and easy solution to prevent the 500 error is to send the
email before redirecting. However, we can go a step further and try to
make the experience better for the user by retrying before timing out.

Note that the AWS documentation is not great, and seems to indicate
that these retries do not apply to SES, but it doesn't hurt to include
the retry options just in case they do work with SES.

I also took the opportunity to refactor the controller and move the
business logic out of it and into the Form Object. With this change,
the extra analytics attributes that check if the user is confirmed
or if they are verified no longer make sense in `ResetPasswordForm`
because they won't reflect the state of the user prior to the password
change. Before, the controller was handling the user updates after the
form returned the analytics attributes. Now that the form is updating
the user before it returns the FormResponse object, the data is no
longer valid.

What we are really interested in here is the state of the user at the
time they request a password reset. We want to see what percentage of
users request a password reset before they have finished creating their
account, and how often verified users request password resets. The point
in time at which it makes the most sense to capture this state is when
they enter their email in the password reset form, which is why those
analytics attributes were moved to `PasswordResetEmailForm`.