LG-691 Increase minimum password length to 12
**Why**:
We have decided to increase the security of the password since we rely
on it for PII encryption as well as the secret to access the account.
After an anonymous analysis of existing users, we found that 50% of
users had already chosen to have a password that is 12 characters or
longer.