Don't use blank padding in event timestamps
**Why**: It causes tests to fail when the day of the month is a single
digit. For example, in the account_history_spec, the event creation
dates change every day because they are set to the current time minus
some arbitrary amount of days. To display the timestamp to the user, we
were using the `%e` option for the integer day of the month, which uses
blank padding. Therefore, a date like `July 4, 2018` would contain two
spaces after July.

In the latest Capybara version, which we upgraded to recently, looking
for content using `have_content` looks at what's being output literally.
In previous Capybara versions, extra whitespace was stripped, which
could mask certain bugs.

**How**: Use the `%-d` option instead, which doesn't use blank padding.