LG-667 Don't create stray ServiceProvideRequests (#2563)
* LG-667 Don't create stray ServiceProvideRequests

**Why**: When we first introduced the `ServiceProvideRequests` table
(https://github.com/18F/identity-idp/pull/1265), it was to allow users
to go through the account creation process across multiple browsers and
still end up back at the SP. This initial implementation did not result
in any stray entries in the `ServiceProvideRequests` table in the
typical sign in scenario.

However, we noticed a bug with this implementation that prevented users
from creating multiple accounts during the same session, so we removed
the guard clause in the SAML and OIDC controllers
(https://github.com/18F/identity-idp/pull/1542), resulting in stray
entries being created, which we acknowledged in the PR commit message
and proposed using a rake task to clean them up, but never followed up.

Fast forward about a year later, and the table now was millions of
entries and is causing DB errors and slowdowns. Upon revisiting the
history of this table, and thinking about the problem some more, I found
a solution that allows us to support all scenarios while also preventing
stray entries. This solution also reduces the total amount of DB
transactions and cuts writes in half.

**Before:**
1. User (who is not signed in) makes request A from SP
2. Request A is stored in the DB and in the session by the controller
(this requires a DB read)
3. User signs in + 2FA
4. ApplicationController looks up Request A from the session to redirect
the user back to the entry point controller
5. The controller stores Request A again as a new entry in the DB,
which we'll call Request B. The info is stored in the session again as
well, resulting in another DB read.
6. User is redirected back to SP, and Request B is deleted from the DB,
but Request A remains in the DB.

Total DB transactions: 2 writes, 2 reads, 1 delete

**After:**
1. User (who is not signed in) makes request A from SP
2. Request A is stored in the DB and in the session by the controller
(this requires a DB read)
3. User signs in + 2FA
4. ApplicationController looks up Request A from the session to redirect
the user back to the entry point controller
5. The controller compares the request stored in the session (DB read)
with the one that was just received by the controller and sees that they
are both from the same SP, and does create a new entry in the DB
6. User is redirected back to SP, and Request A is deleted from the DB.

Total DB transactions: 1 write, 2 reads, 1 delete

**How**: Instead of using the presence of the request_id in the session
as the guard clause, compare the SP that made the last request with
the one stored in the session. If they're the same, don't store a
new entry in the DB, otherwise, delete the entry that matches the one
in the session, and then create a new entry.