[LG-589] read email from email_address table (#2570)
* [LG-589] read email from email_address table

**Why**:
We would like to support multiple email addresses per user.
But first, we need to start using the email_address table without
having to worry yet about *which* email address we're using since
we're limiting it to one for now. Future PRs will open this up
to multiple and add support for selecting email for SPs, etc.

**How**:
Move from reading `email` and `confirmed?` from the `user` object
to the `user.email_address` object.