LG-688 Fix Account Reset Notifications analytics
**Why**: While looking for Account Reset events in Kibana and Redshift,
I noticed the `notifications` type of events were missing. I didn't
think the issue was that the AWS Lambda wasn't running because that
would mean no one would be able to delete their accounts, yet I saw
events for tokens granted and users completing the deletion. I also
confirmed this by seeing calls being made to the notifications
controller endpoint in the Rails production logs. I also tested locally
by running the lambda directly and by sending a cURL request, and saw
the events being recorded.

Then I remembered that the Ahoy gem excludes events if it thinks the
user agent is a bot, but that didn't explain why it would work locally.
It turns out that we have a guard clause in the Ahoy `exclude?` method
that doesn't attempt to exclude any requests if
`FeatureManagement.use_dashboard_service_providers?` is `true`, which
it is by default in development. After setting the config to `false`,
I noticed that the event was no longer recorded.

This led me to run `bundle open ahoy` to read the gem's code and I
noticed that they switched to using the `DeviceDetector` gem for user
agent parsing earlier this year (which is our preferred parser
as well that we use in the Analytics class), but we missed the memo
when upgrading from 2.0 to 2.1. In order to use `DeviceDetector`, we
need to set it explicitly in the initializer. Otherwise, Ahoy will use
the `Browser` gem.

The bug is that the `Browser` gem thinks the request to the
notifications endpoint is from a bot, whereas `DeviceDetector` does not.