Fix multiple MFA configurations login options bug
**Why**:
When a user has multiples of certain MFA configurations, we only want
to display the option once during login. For example, when a user
configures multiple Webauthn keys, either one will work, so we only
need to show the "Hardware security key" option once in the list of
security options displayed during login.

However, when a user has multiple MFA-enabled phone numbers, we want
to show them all so that the user can choose one in case they lost
access to one of them, for example.

The code in #2576 did not fix the issue with showing multiple
"Hardware security key" radio buttons. The feature specs added in this
PR fail on the master branch, and I confirmed it while testing locally.
I don't know why the presenter specs passed.

**How**:
- Use plain Ruby Enumerable methods to modify the presenters list so
that it only keeps one `WebauthnSelectionPresenter` instance, but allows
multiples of Sms and Voice.
- Remove unnecessary `WebauthnConfigurationsExtension` and related code.
- Update `PhoneSelectionPresenter#type` to use underscore instead of
colon to stay consistent with `id` names (Capybara doesn't like the
colon).
- Only show the last 4 digits of the phone number in the login options.