Only show webauthn option at most once
**Why**:
We send all the webauthn configurations to the browser when
a user is using FIDO, so it doesn't make sense to have multiple
entries in the selection list.

**How**:
Special case the webauthn configurations list so we call the
extension to get the selection presenter for the collection
rather than the individual configurations.