LG-721 Track account age and enabled MFA methods (#2574)
* LG-721 Track account age and enabled MFA methods

**Why**:
Since the account deletion feature deletes a user's account, we cannot
look up some of their attributes after the fact. Some attributes that
would be useful to capture before the account is deleted are:

- The account age, i.e. the difference in time between the time of
deletion request and the account confirmation time (note that this can
be obtained from analytics data, but having the number right there in
the event is more convenient)

- Which MFA methods are configured

This will allow us to measure the current state of affairs, and then
compare it after we introduce the ability to add backup phone numbers.
For example, if one of our goals is to reduce the amount of account
deletion requests, and if we notice that the average amount of time it
takes for users to lose access to their phone is 1 month, then we know
we should wait at least a month before looking at the new numbers.

Similarly, if most account deletion requests are currently made by folks
who only have 1 phone as their MFA method, and if a month after we
deploy backup phone support the number of deletion requests from folks
with only 1 phone as their MFA method doesn't go down, it could mean
that we need to improve the way we are making users aware of the
feature, or that perhaps backup phone numbers is the wrong solution
because most folks who lose access to their devices don't have a backup
phone number to provide.