Add full email_fingerprint index
**Why**:
The current email fingerprint index is for confirmed
accounts and forces the fingerprints to be unique for
confirmed accounts. This left unconfirmed accounts out
in the cold. Which was fine as long as we read from the
users table. Now we're reading from the email_addresses
table.

**How**:
Add an index on email_fingerprint that doesn't require
uniqueness.