Use remember_me_revoked_at to revoke remember device
**Why**: To decouple the remember device action from the phone
configurations and make revoking remember device sessions an explicit
action rather than a side effect of updating the phone configuration

To be able to deploy this change, it needs to be spread over 2 releases.
One to add and write to the new column and a later one to read from it.
This PR also includes a rake task to backfill this column on the
existing user table.