LG-359 Don't penalize users for IdV vendor errors
**Why**:
Previously, if a verification failure occurred due to a network error or
system error, it was treated the same as an issue with the data
provided by the user, thereby decrementing the user's remaining
attempts. The limit on attempts is meant to prevent bad actors from
fishing for data, not to penalize the user for technical issues that are
out of their control.

**How**:
Don't increment the attempts count if the errors result from non-data
related verification failures, such as vendor exceptions and timeouts.